Implement view-source scheme for web requests, and only web requests.
To determine whether to return the actual source, or not, I added a boolean to webrequests.
When the Url tries to show the request content, it now checks to see if it's both a webRequest and the viewSource flag is shown; if so, it just prints the body, and if both aren't true, it prints the stuff like normal.